Skyblock-AH-flipper new version
This is a working ah flipper for hypixel skyblock that if used correctly can make u a lot of coins. once you downoad the mod you should put it on your mod folder and run it in forge 1.8.9. if you have any questions on how to use it pls dm me on discord Toxic_Oficial#9608

-fixed some bugs in this new release